Add mage target for go benchmarks and benchstat
This commit add 2 new mage targets for running the go naitive benchmark
tests and running the benchstat for parsing the results.

With the following targets you can do the following

- Run benchmarks and print or save the output
```bash
$ mage benchmark:run
```

This will execute the go benchmark.
You can control the number of iterations that the benchmark will execute
by defining an env var `BENCH_COUNT`
You can write the output into a file by defining the `OUTPUT` variables.

Example:
```bash
$ OUTPUT=base.txt BENCH_COUNT=1 mage benchmark:run
```

When it finished there is going to be a file `base.txt` with the output
of the executed benchmarks.

- Examining the benchmark output

```bash
$ BASE=base.txt mage benchmark:diff
```

This will parse and print the output of the benchmark.

To take the most out of the diff is when comparing benchmark results.

Assuming that you have executed the `benchmark:run` before you start
performing your changes and have the output stored into the `base.txt`
then after you have completed your changes you can re-run the same
benchmark command, but this time write it to `next.txt`.
Then you can compare the outcome by the following command.

```bash
$ BASE=base.txt NEXT=next.txt mage benchmark:diff
```

Defining an OUTPUT variable in the above will save the comparison in the
defined output file.

Signed-off-by: Alexandros, Sapranidis <alexandros@elastic.co>