monitoring - Rename NewRegistry to MustNewRegistry
registry.NewRegistry could panic and it was not documented. This is a breaking
change to force callers of this method to evaluate if they are willing to accept the
current behavior which panics.

This creates two choices for existing users. The can use either:

    MustNewRegistry(...) *Registry
    NewRegistry(...) (*Registry, error)

A similar change was made to the `registry.Add` method (except existing code will
continue to compile since it previously had no return value).