Add Close method to logp.Logger (#205)
Add a close method to logp.Logger responsible for closing any
underlying core/output. All of our implementations of zapcore.Core
that use an output that can be closed now implement io.Closer.

This allows for closing any underlying files used by any output, which
can prevent issues on Windows.

It also removes a workaround for windows in a test because the log files
could not be closed.