logp: add rate limited logger
This commit adds support for rate limiting to logp.Logger. This is useful
when you have a lot of log messages and you want to avoid flooding the
destination.

The logger was augmented with the following methods:
    `Throttled` will only log once every period.
    `Sampled` will only log the nth message.
    `Limited` will only log the first n messages.

The implementation is based on x/time/rate.Sometimes.

For all these limiters any message that is not logged will be dropped.

These also synergize with one another, for example:

    log.Sampled(10).
        Throttled(time.Second).
        Info("")`

will log one message every second, and only one out of ten messages.

Likewise,

    log.Limited(10).
        Throttled(time.Second).
        Info("")

will log the first ten messages, and then one message every second.