Add a passphrase_file config option when creating keystores
Add a passphrase_file config option to allow users of beats to specify a
password that exists outside of the beat.yml that is used as a keystore
password. Having a password is a requirement for our FIPS distributions.
When a binary is built without the requirefips tag no behaviour changes,
but if the tag is used a passphrase_file is required. This file must
exist and be non-empty.