config: protect global configOpts with a mutex
There is a race condition in configOpts when one goroutine calls
OverwriteConfigOpts while another reads the variable. This is common, as
almost every method in config.C reads this global variable.

Fix the data race by using a mutex. Since this variable is mostly read,
optimize it with a sync.RWMutex.