Enable GODEBUG=fips140=only tests in buildkite.
Enable unit tests with GODEBUG=fips140=only and -tags=requirefips in
buildkite. This is a basic check to make sure we don't accidently add
non-compliant crypto when in FIPS mode.
When these test are ran, the encrypted private key tests in
transport/tlscommon are skipped, the -tags=requirefips tests check that
an error is returned, however if fips140=only is set the test will
panic.
Add a new V2 file keystore that is used with requirefips. This
implementation uses NewGCMWithRandomNonce. File keystore support is
currently disabled in FIPS artifacts we produce.