Add helper to create a sub-registry or return the existing one if present (#313)
When creating a child registry, code often checks if the registry already exists, since creating a duplicate entry will cause a panic. There are two issues with this:

- Most registry creation has extra identical boilerplate to check if a registry exists before creating it
- This boilerplate is not concurrency-safe, because the registry lock is not held continuously; it's possible for two goroutines to both check the same registry name, both observe that it does not exist yet, and both try to create it. This is rare (I don't know of confirmed occurrences) but if it does happen the entire process will panic.

This PR adds a helper that handles both these issues, replacing the usual checks with a one-line call that properly holds the lock between checking for existing registries and creating new ones.