Prevent logptest.NewTestingLogger from panics.
Because NewTestingLogger used testing.T as the logger output, if the
logger is used after the test has ended, it will panic.

This behaviour has been seen a few times and causes flakiness in our
CI.

To mitigate this, NewTestingLogger now just calls NewFileLogger that
uses a file as the logger output, thus being safe for use even after
the test has ended.